+++ src/sync/events/index.ts
@@ -1763,6 +1763,7 @@ export const syncEvents = async (
               }
 
               const orderKind = "rarible";
+              const orderSource = await getOrderSourceByOrderKind(orderKind);
 
               let taker = rightMaker;
 
@@ -1781,6 +1782,7 @@ export const syncEvents = async (
                 orderKind: "rarible",
                 orderId: leftHash,
                 orderSide: side,
+                orderSourceIdInt: orderSource?.id,
                 maker: leftMaker,
                 taker,
                 price: prices.nativePrice,
@@ -1820,10 +1822,14 @@ export const syncEvents = async (
                 break;
               }
 
+              const orderKind = "element-erc721";
+              const orderSource = await getOrderSourceByOrderKind(orderKind);
+
               fillEventsPartial.push({
-                orderKind: "element-erc721",
+                orderKind,
                 orderId: orderHash,
                 orderSide: "sell",
+                orderSourceIdInt: orderSource?.id,
                 maker,
                 taker,
                 price: prices.nativePrice,
@@ -1863,10 +1869,14 @@ export const syncEvents = async (
                 break;
               }
 
+              const orderKind = "element-erc721";
+              const orderSource = await getOrderSourceByOrderKind(orderKind);
+
               fillEventsPartial.push({
-                orderKind: "element-erc721",
+                orderKind,
                 orderId: orderHash,
                 orderSide: "buy",
+                orderSourceIdInt: orderSource?.id,
                 maker,
                 taker,
                 price: prices.nativePrice,
@@ -1907,10 +1917,14 @@ export const syncEvents = async (
                 break;
               }
 
+              const orderKind = "element-erc1155";
+              const orderSource = await getOrderSourceByOrderKind(orderKind);
+
               fillEventsPartial.push({
-                orderKind: "element-erc1155",
+                orderKind,
                 orderId: orderHash,
                 orderSide: "sell",
+                orderSourceIdInt: orderSource?.id,
                 maker,
                 taker,
                 price: prices.nativePrice,
@@ -1951,10 +1965,14 @@ export const syncEvents = async (
                 break;
               }
 
+              const orderKind = "element-erc1155";
+              const orderSource = await getOrderSourceByOrderKind(orderKind);
+
               fillEventsPartial.push({
-                orderKind: "element-erc1155",
+                orderKind,
                 orderId: orderHash,
                 orderSide: "buy",
+                orderSourceIdInt: orderSource?.id,
                 maker,
                 taker,
                 price: prices.nativePrice,
@@ -2070,14 +2088,28 @@ export const syncEvents = async (
               const winner = args["winner"].toLowerCase();
               const amount = args["amount"].toString();
 
+              const currency = Sdk.Common.Addresses.Eth[config.chainId];
+
+              const prices = await getPrices(currency, amount, baseEventParams.timestamp);
+
+              if (!prices.nativePrice) {
+                // We must always have the native price
+                break;
+              }
+
+              const orderKind = "nouns";
+              const orderSource = await getOrderSourceByOrderKind(orderKind);
+
               fillEvents.push({
-                orderKind: "nouns",
+                orderKind,
+                orderSourceIdInt: orderSource?.id,
                 orderSide: "sell",
                 maker: Sdk.Nouns.Addresses.AuctionHouse[config.chainId]?.toLowerCase(),
                 taker: winner,
                 amount: "1",
-                currency: Sdk.Common.Addresses.Eth[config.chainId],
-                price: amount,
+                currency,
+                price: prices.nativePrice,
+                usdPrice: prices.usdPrice,
                 contract: Sdk.Nouns.Addresses.TokenContract[config.chainId]?.toLowerCase(),
                 tokenId: nounId,
                 baseEventParams,